renamed online_request_processor -> base_online_request_processor
